parse, gen and apply together as one function
